refactor: Move the pages type
