+++ constants/index.ts
@@ -1,5 +1,5 @@
 export { COOKIES } from './cookies'
-export { type Pages, CRAWLABLE, HANDLES, LINKS, SOCIALS } from './links'
+export { CRAWLABLE, HANDLES, LINKS, SOCIALS } from './links'
 export {
   CONTENTTYPE,
   METAFONT,
