+++ scripts/mdx.ts
@@ -10,8 +10,8 @@ import rehypeMdxCodeProps from 'rehype-mdx-code-props'
 import rehypeUnwrapImages from 'rehype-unwrap-images'
 import { Pluggable } from 'unified'
 import readingTime from 'reading-time'
-import type { Locale, MDX, MDXTypes, Post, Project } from '@/types'
-import { LINKS, Pages } from '@/constants'
+import type { Locale, MDX, MDXTypes, Pages, Post, Project } from '@/types'
+import { LINKS } from '@/constants'
 import { localizeFileName, localizePath } from '@/lib'
 import { getMDXComponents } from '@/mdx-components'
